+++ example/lib/main.mapper.g.dart
@@ -53,7 +53,20 @@ class PersonMapper extends BaseMapper<Person> {
 extension PersonMapperExtension on Person {
   String toJson() => Mapper.toJson(this);
   Map<String, dynamic> toMap() => Mapper.toMap(this);
-  Person copyWith({String? name, int? age, Car? car}) => Person(name ?? this.name, age: age ?? this.age, car: car ?? this.car);
+  PersonCopyWith<Person> get copyWith => PersonCopyWith(this, _$identity);
+}
+
+abstract class PersonCopyWith<$R> {
+  factory PersonCopyWith(Person value, Then<Person, $R> then) = _PersonCopyWithImpl<$R>;
+  CarCopyWith<$R>? get car;
+  $R call({String? name, int? age, Car? car});
+}
+
+class _PersonCopyWithImpl<$R> extends BaseCopyWith<Person, $R> implements PersonCopyWith<$R> {
+  _PersonCopyWithImpl(Person value, Then<Person, $R> then) : super(value, then);
+
+  CarCopyWith<$R>? get car => _value.car != null ? CarCopyWith(_value.car!, (v) => call(car: v)) : null;
+  $R call({String? name, int? age, Object? car = _none}) => _then(Person(name ?? _value.name, age: age ?? _value.age, car: or(car, _value.car)));
 }
 
 class CarMapper extends BaseMapper<Car> {
@@ -77,7 +90,18 @@ class CarMapper extends BaseMapper<Car> {
 extension CarMapperExtension on Car {
   String toJson() => Mapper.toJson(this);
   Map<String, dynamic> toMap() => Mapper.toMap(this);
-  Car copyWith({int? drivenKm, Brand? brand}) => Car(drivenKm ?? this.drivenKm, brand ?? this.brand);
+  CarCopyWith<Car> get copyWith => CarCopyWith(this, _$identity);
+}
+
+abstract class CarCopyWith<$R> {
+  factory CarCopyWith(Car value, Then<Car, $R> then) = _CarCopyWithImpl<$R>;
+  $R call({int? drivenKm, Brand? brand});
+}
+
+class _CarCopyWithImpl<$R> extends BaseCopyWith<Car, $R> implements CarCopyWith<$R> {
+  _CarCopyWithImpl(Car value, Then<Car, $R> then) : super(value, then);
+
+  $R call({int? drivenKm, Brand? brand}) => _then(Car(drivenKm ?? _value.drivenKm, brand ?? _value.brand));
 }
 
 class BoxMapper extends BaseMapper<Box> {
@@ -101,7 +125,18 @@ class BoxMapper extends BaseMapper<Box> {
 extension BoxMapperExtension<T> on Box<T> {
   String toJson() => Mapper.toJson(this);
   Map<String, dynamic> toMap() => Mapper.toMap(this);
-  Box<T> copyWith({int? size, T? content}) => Box(size ?? this.size, content: content ?? this.content);
+  BoxCopyWith<Box<T>, T> get copyWith => BoxCopyWith(this, _$identity);
+}
+
+abstract class BoxCopyWith<$R, T> {
+  factory BoxCopyWith(Box<T> value, Then<Box<T>, $R> then) = _BoxCopyWithImpl<$R, T>;
+  $R call({int? size, T? content});
+}
+
+class _BoxCopyWithImpl<$R, T> extends BaseCopyWith<Box<T>, $R> implements BoxCopyWith<$R, T> {
+  _BoxCopyWithImpl(Box<T> value, Then<Box<T>, $R> then) : super(value, then);
+
+  $R call({int? size, T? content}) => _then(Box(size ?? _value.size, content: content ?? _value.content));
 }
 
 class ConfettiMapper extends BaseMapper<Confetti> {
@@ -125,7 +160,18 @@ class ConfettiMapper extends BaseMapper<Confetti> {
 extension ConfettiMapperExtension on Confetti {
   String toJson() => Mapper.toJson(this);
   Map<String, dynamic> toMap() => Mapper.toMap(this);
-  Confetti copyWith({String? color}) => Confetti(color ?? this.color);
+  ConfettiCopyWith<Confetti> get copyWith => ConfettiCopyWith(this, _$identity);
+}
+
+abstract class ConfettiCopyWith<$R> {
+  factory ConfettiCopyWith(Confetti value, Then<Confetti, $R> then) = _ConfettiCopyWithImpl<$R>;
+  $R call({String? color});
+}
+
+class _ConfettiCopyWithImpl<$R> extends BaseCopyWith<Confetti, $R> implements ConfettiCopyWith<$R> {
+  _ConfettiCopyWithImpl(Confetti value, Then<Confetti, $R> then) : super(value, then);
+
+  $R call({String? color}) => _then(Confetti(color ?? _value.color));
 }
 
 
@@ -452,3 +498,19 @@ extension MapGet on Map<String, dynamic> {
     }
   }
 }
+
+class _None { const _None(); }
+const _none = _None();
+
+T _$identity<T>(T value) => value;
+typedef Then<$T, $R> = $R Function($T);
+
+class BaseCopyWith<$T, $R> {
+  BaseCopyWith(this._value, this._then);
+
+  final $T _value;
+  final Then<$T, $R> _then;
+  
+  T or<T>(Object? _v, T v) => _v == _none ? v : _v as T;
+}
+