+++ lib/src/utils.dart
@@ -1,9 +1,13 @@
 import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/nullability_suffix.dart';
+import 'package:analyzer/dart/element/type.dart';
 import 'package:source_gen/source_gen.dart';
 
 import 'core/annotations.dart';
+import 'core/case_style.dart';
 import 'core/mappers.dart';
 
 const enumChecker = TypeChecker.fromRuntime(MappableEnum);
@@ -82,4 +86,18 @@ Iterable<ClassElement> elementsOf(LibraryElement element) sync* {
 
 extension OptIter<T> on Iterable<T> {
   T? get firstOrNull => isEmpty ? null : first;
+}
+
+CaseStyle? caseStyleFromAnnotation(DartObject obj) {
+  return CaseStyle(
+    head:
+        TextTransformParser.parse(obj.getField('head')!.toStringValue() ?? ''),
+    tail:
+        TextTransformParser.parse(obj.getField('tail')!.toStringValue() ?? ''),
+    separator: obj.getField('separator')!.toStringValue() ?? '',
+  );
+}
+
+extension NullableType on DartType {
+  bool get isNullable => nullabilitySuffix == NullabilitySuffix.question;
 }